+++ package.json
@@ -1,10 +1,10 @@
 {
     "scripts": {
-        "start": "webpack-dev-server",
+        "start": "webpack-dev-server --quiet",
         "build": "webpack"
     },
     "dependencies": {
-        "@lucsoft/webgen": "^1.6.12",
+        "@lucsoft/webgen": "^1.6.15",
         "html-webpack-plugin": "^3.2.0",
         "ts-loader": "^6.2.1"
     },
