+++ src/modules/dashboard/index.ts
@@ -1,8 +1,9 @@
-import { DataConnect } from '@lucsoft/webgen';
-import { ElementResponse } from '@lucsoft/webgen/bin/classes/ElementsResponse';
+import { DataConnect, ElementResponse } from '@lucsoft/webgen';
 
 import { page, web } from '../app';
 import { HomeSYSModule } from '../app/modules';
+import { FightOfLife } from './FightOfLife';
+import { HomeSYSAppModule } from './modules';
 
 class cardbutton
 {
@@ -125,188 +126,59 @@ export class DashboardModule extends HomeSYSModule
                 }
             ]
         }).next.pageTitle({
-            text: "HomeSYS Connect"
+            text: "Connected Systems"
         }).modify.element.style.marginTop = "4rem";
-        if (!data.profile.modules.homesys.homsysConnectConnected)
-        {
-            this.cards.next.cards({
-                small: false,
-                columns: "1",
-                hidden: false,
-                cards: [
-                    {
-                        title: "Unable to Connect",
-                        subtitle: "HomeSYS Connnect has not logged into HmSYS",
-                        id: "homesysconnect"
-                    }
-                ]
-            })
-        }
+
+        const hRendering = require('./ConnectedSystems/houseRendering').HouseRendering;
+        new hRendering(this.cards, data);
+
         setInterval(() =>
         {
             document.querySelector('#hmsysrunningsince').querySelector('.title').innerHTML = timeAgo(data.profile.modules.homesys.runningSince);
         }, 1000);
-        this.cards.next.pageTitle({
-            text: "Online Modules"
-        }).modify.element.style.marginTop = "4rem";
+        const list = [];
+        const hmsys = data.profile.modules.hmsys;
+        if (data.profile.modules.hmsys == undefined)
+            return;
+
+        var moduleList = [
+            {
+                title: "Global Settings",
+                value: " ",
+                id: "homeSYSSettings"
+            },
+            {
+                title: "FightOfLife",
+                subtitle: "Proof of Concept Game",
+                id: "FightOfLife"
+            }
+        ];
+
+        for (const mod of moduleList)
+        {
+            if (hmsys.loadedModules.includes(mod.id))
+                list.push({
+                    title: mod.title,
+                    value: mod.subtitle,
+                    id: mod.id,
+                    onClick: async () =>
+                    {
+                        const loadedModule: typeof HomeSYSAppModule = (await this.loadModule(mod.id))[ mod.id ];
+                        new loadedModule(web, this.cards, data);
+                    }
+                })
+        }
+
         this.cards.next.cardButtons({
             small: false,
             columns: "3",
-            list: [
-                {
-                    title: "HomeSYS Settings",
-                    value: "Global",
-                    id: "homsys",
-                    onClick: () =>
-                    {
-                        this.openSettings();
-                    }
-                },
-                {
-                    title: "FightOfLife",
-                    value: "Proof of Concept",
-                    id: "fol",
-                    onClick: () =>
-                    {
-                        this.openGame();
-                    }
-                }
-            ]
+            list: list
         }).modify.element.style.marginBottom = "5rem";
 
     }
 
-    private openGame()
+    private async loadModule(moduleID: string)
     {
-        const game = web.elements.layout("fixedWindow").element;
-        game.element.onclick = (ev: any) =>
-        {
-            if (ev.path[ 0 ].nodeName == "ARTICLE")
-            {
-                // game.element.remove();
-                // document.body.style.overflow = "unset";
-            }
-        }
-        var login = web.elements.add(game.element).login({
-            email: "Gabe Newell",
-            button: "Join eu01",
-            text: "What's your name?",
-            login: (_, email) =>
-            {
-                console.log(this.data);
-                this.data.triggerCommand("registerCLMP", { username: email });
-                login.modify.element.remove();
-                this.gameLogic(login);
-            }
-        });
-
-    }
-
-    private gameLogic(windowElement: ElementResponse)
-    {
-        document.body.style.overflow = "hidden";
-        var gameBlock = document.createElement('div');
-        var canvas = document.createElement('canvas') as HTMLCanvasElement;
-        var game = canvas.getContext("2d");
-        game.fillStyle = "#FFFFFF";
-        game.fillRect(0, 0, 3000, 2000);
-        game.fillStyle = "#000000";
-        for (let index = 0; index < 40; index++)
-        {
-            game.fillRect(10 + index, 10 + index, 1, 1);
-        }
-        canvas.style.imageRendering = "pixelated";
-        canvas.style.height = "2000px";
-        canvas.style.width = "3000px";
-
-        gameBlock.append(canvas);
-        gameBlock.style.overflow = "auto";
-        gameBlock.style.overflowX = "hidden";
-        gameBlock.style.height = "20rem";
-        gameBlock.style.width = "35rem";
-        gameBlock.style.margin = "0 auto";
-        gameBlock.scrollTop = 1;
-        var currentScale = 1.2;
-        gameBlock.style.position = "relative";
-        canvas.style.position = "relative";
-
-        canvas.style.transform = `scale(${currentScale})`;
-        gameBlock.addEventListener('scroll', (e) =>
-        {
-            if (gameBlock.scrollTop == 0)
-            {
-                currentScale += 0.1;
-                gameBlock.scrollTop = 1;
-            } else if (gameBlock.scrollTop > 1)
-            {
-                if (currentScale > 0.11)
-                    currentScale -= 0.1;
-                gameBlock.scrollTop = 1;
-            }
-            this.changeTransfrom(canvas, currentScale)
-        });
-
-        var isDown = false;
-        var offset = [ 0, 0 ];
-        var mousePosition;
-        canvas.addEventListener('mousedown', (e) =>
-        {
-            isDown = true;
-            offset = [
-                canvas.offsetLeft - e.clientX,
-                canvas.offsetTop - e.clientY
-            ];
-        }, true);
-
-        document.addEventListener('mouseup', () =>
-        {
-            isDown = false;
-        }, true);
-
-        document.addEventListener('mousemove', (event) =>
-        {
-            event.preventDefault();
-            if (isDown)
-            {
-                mousePosition = {
-
-                    x: event.clientX,
-                    y: event.clientY
-
-                };
-
-                canvas.style.left = (mousePosition.x + offset[ 0 ]) + 'px';
-                canvas.style.top = (mousePosition.y + offset[ 1 ]) + 'px';
-
-            }
-        }, true);
-        windowElement.next.window({
-            title: "FightOfLife playling on eu01.hmsys.de",
-            content: gameBlock,
-            maxWidth: "40rem"
-        });
-    }
-
-    private changeTransfrom(canvas: HTMLCanvasElement, currentScale: number, )
-    {
-        canvas.style.transform = `scale(${currentScale})`;
-    }
-    private openSettings()
-    {
-        const settings = web.elements.layout("fixedWindow").element;
-        settings.element.onclick = (ev: any) =>
-        {
-            if (ev.path[ 0 ].nodeName == "ARTICLE")
-            {
-                settings.element.remove();
-            }
-        }
-        var test = document.createElement('button');
-        test.onclick = () => web.elements.layout("fixedWindow", true);
-        test.innerHTML = "Close";
-        web.elements.add(settings.element).window({
-            title: "Settings",
-            content: [ `NEEE`, test ],
-        });
+        return eval(await (await fetch(`./module/${moduleID}.js`)).text());
     }
 }
